requirements.txt
opencv-python
Wand
black
isort
pylint
pre-commit
pytest
fastapi
uvicorn[standard]
typing
python-multipart
pydantic
numpy
requests
Pillow
scipy
